pyproject.toml
[project]
name = "yt-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.23",
    "djangorestframework>=3.16.0",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "yt-dlp>=2025.6.30",
]
